requirements.txt
datasets==2.12.0
faiss-cpu==1.7.4
gitdb==4.0.10
numpy==1.24.3
pandas==2.0.2
Pillow==9.5.0
pyarrow==12.0.0
streamlit==1.23.1